cambio de estructura de filtros de busqueda
